UPDATE: Ended EventService with his HandleEvent and HandleHistoryRecord mathods.Fixed some dtos. Tested to run, Swagger is opening and all works well, but can't test further

+++ Models/Dtos/Bitrix/CallInfoDto.cs
@@ -4,7 +4,7 @@
     {
         public string CallId { get; init; }
         public string? UserPhoneInner { get; init; }
-        public int? UserId { get; init; }
+        public long? UserId { get; init; }
         public int Duration { get; init; } = -1;
         public double? Cost{ get; init; }
         public string? CostCurrency { get; init; }
